Move filter section on works .js to filter.js
