+++ index.html
@@ -12,9 +12,9 @@
     />
     <meta name="description" content="" />
     <link rel="stylesheet" href="./assets/style.css" />
-    <script type="module" src="works.js" defer></script>
-    <script src="filter.js" defer></script>
     <script type="module" src="authentification.js" defer></script>
+    <script type="module" src="works.js" defer></script>
+    <script type="module" src="filter.js" defer></script>
     <script type="module" src="worksManager.js" defer></script>
   </head>
   <body>
